requirements.txt
# Django/Framework Packages
django>=1.4.12
django-model-utils==1.4.0
South>=0.7.6

# Testing Packages
coverage==3.7.1
django_nose
nose==1.3.3
httpretty==0.8.0
pep8==1.5.7
pylint==1.2.1
pep257==0.3.2
mock==1.0.1
testfixtures==4.0.0

# edX Packages
-e git+https://github.com/edx/opaque-keys.git@cba8e23f53ff3ce9d1f45871986c8974576185c0#egg=opaque-keys
